Make reader re-creation possible
As it was, the current code made it hard to re-create a reader for the
data we run our benchmark against. This change moves the creation of the
reader object into a small inner function that we now call in a with
block to ensure it's always closed.

Moreover, the reader object has always been created, so there was some
code checking for its existence that shouldn't have been there in the
first place.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>