Bump Python version used to build sources and binaries
3.7 is EOL, and therefore we're not getting the latest version of
cibuildwheels, since they stopped supporting Python 3.7 a while ago.
This in turn means we are not getting wheels for Python 3.12.

In this commit we're updating our actions to use 3.12 to get the longest
lifetime possible. In the case of the source distribution, we
additionally need to install setuptools since distutils isn't included
in CPython anymore.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>