Remove Python code for compatibility with < 3.5
Most of the compat module dealt with the differences between Python 2
and 3, and now that we don't formally support 2.7 anymore we can finally
remove it. Note that the module itself still contains code for a
different type of compatibility (between strings and bytes), which we
still need to retain.

Additionally, the module defined a flag that helped identifying when the
code was running under 3.5+. This is now always the case, which results
in some simplifications, most notably unconditional import of the
utils35 module, and the ability to define coroutines freely.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>