Reorganise package and extension sources
Two main changes are taking place: first, all sources are now put under
src/, making it possible to test the C extension backend when wheels are
installed, even if tests are run directly from the top-level directory,
like cibuildwheel does. Previously this wouldn't work, because the local
"ijson/" subfolder would take precedence, and there was no _yajl2 module
present there.

Secondly, the directory containing the extension has been renamed from
yajl2_c to ext/_yajl2. The old name wasn't really correct, and putting
things under an ext/ subdirectory makes it easier for the casual reader
to realise there is a C extension within.

As a side effect of correctly indicating where the "ijson" package is in
the source tree, we have now eliminated the packageing of the "tests"
directory, which was previously considered a package. The directory is
now not inside the src/ directory, and therefore not found by
find_packages. In order to still have it included in the source
distributions we then need to graft it (only tests/test*.py files are
automatically included, and we still have it structured as a package) in
the MANIFEST.in file.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>