Skip wheel building in old PyPy versions
Versions of PyPy < 7.3.13 crashed due to a bug on how they emulated some
of the Python C API. It thus doesn't make sense to build binary wheels
for version of Python that were not supported starting from PyPy 7.3.13
(we still want to run tests against PyPy though).

More background and pointers in #101.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>