Inject backend into main entry point tests
Using the new "backend" automatic parametrisation implemented in the
previous commit we can now remove the dynamic class generation mechanism
previously used and replace it with receiving the new "backend" fixture.
As a side-effect of not using the dynamic class creation logic anymore,
we need to rename our test class to being with Test so it gets picked up
by pytest.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>