Handle creation errors correctly in async generators
Our async generators in the yajl2_c extension internally create other
objects, some of which might raise exceptions. These were not correctly
handled though, and we were blindly swallowing them, but not returning
NULL.

This commit adds a return value to the async_reading_generator_add_coro
function that internally creates the rest of the objects used by the
async generators, returning 0 for success and -1 in the case of a
failure.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>