Set exception if buf_size is not a number
We were previously doing the check, but incorrectly returning -1 without
setting an exception, probably under the wrong assumption that
PyNumber_Check would do it for us.

We also need to be careful to perform the check after we increment the
reference count for the objects we received via PyArg_ParseTuple.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>