Introduce yajl_parse_context structure
We have historically used the target_send object as the context provided
to yajl for its parsing logic. While this works, it's not obvious and
doesn't allow fur further expansion in the future.

This commit introduces as new yajl_parse_context structure to use as the
parsing context for yajl. It currently contains the target_send
function. The code needing to access the target_send object needs now to
get it from the context.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>