Port to PEP 489 multi-phase module initialisation
Using multi-phase initialisation (and per-module state, yet to come)
will allow us to support Python sub-interpreters, and the free-threading
version of Python, which are the two main ways people might try to
achieve parallelism in newer versions of python.

Additionally a new MODULE_NAME macro keeps the name of the module in a
central place to avoid typos in the future.

Note that this commit doesn't indicate sub-interpreter or free-threading
support, since that requires further changes.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>