Introduce module state
C extensions can have a per-module state object stored in them by
specifying a positive m_size member in their definition. Using a
per-module state is one of the requirements for supporting
sub-interpreters, since each interpreter carries its own module
instance.

The module state structure we use for our C extension contains all the
variables that were previously globally defined, such as event name
singletons and certain imported objects.

This commit simply introduces the module state structure, its
initialisation as part of the module definition (by specifying a
positive m_size that can hold the state), and pointers to it from the
internal object types where we'll require it. This commit does *not*
switch the extension from referring to the globals to using this state,
as that will come in a different commit.

Note that to be able to refer to the module state from our internal
object's methods requires us to internally import our own module to
access the state (which we do as a relative import preferably). This is
because until Python 3.9 it was impossible to access the module object
from an object instance (or its type). Since we want to still support
versions earlier than that we therefore need to access the module level
using this technique, and then cache it in the object instance itself
for further usage when invoking its methods.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>