Use module state instead of global variables
This commits uses the new module state, removing the need for global
variables, and thus making the C extension ready for sub-interpreter
use.

When the C extension is loaded, the module state members are now
initialised instead of the global variables, which have been removed.
Later on, the code that previously read such variables now reads the
information from the module state, which is already available where
required. The bulk of the changes are around the event name singletons
used throughout the code, but there is a couple of imports that we cache
as well.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>