Declare support for sub-interpreters
Now that all pieces are in place, we can confidently declare support for
sub-interpreters (with per-interpreter GIL, since we don't share state
across module instances).

To make sure this is actually true we have a new test that (in a very
fragile manner) checks that ijson can be imported and used in
sub-interpreters, and that we can see the results of those computations
back in the main interpreter.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>