Advertise support for free-threading mode
Python 3.13 ships with an optional free-threading mode where the GIL
isn't held anymore. As far as we can tell, our extension didn't rely on
the GIL for any kind of synchronisation, and all our state (all
read-only) in kept in the module itself, leaving us with no global data
to worry about. Thus, it seems reasonable to advertise our extension as
being able to correctly function in free-threaded mode.

This commit adds the necessary slot to the module definition so the
interpreter keeps the GIL disabled after loading our extension.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>