Replace mktemp by NamedTemporaryFile
We can't use the automatic delete-on-close facilities though because we
need the file to still exist after we finish writing into it, while
being able to open it from a different process. Achieving this without
the new 3.12+ delete_on_close flag doesn't seem possible, so we better
delete it by hand.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>