Keep ctypes' function types alive while in use
We have historically stored these function types in the global space of
the _yajl2_ctypes_common module, but the ctypes documentation clearly
states that users must ensure a reference to them is kept while
callbacks are still issued from C. This is usually true in the common
use case, but it could be a problem at interpreter shutdown situations
where a separate thread is still doing yajl parsing.

This commit fixes this potential issue by keeping the collection of
function types in a single tuple object that is returned alongside the
callbacks themselves every time a new basic_parse routine starts its
execution. This reference is then held by the caller until basic_parse
finishes, thus ensuring the function types are referenced for the full
duration of the yajl parsing.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>