Use new "enable" CIBW option
The "free_threaded_support" option (and related envvar) has been
deprecated in favour of the more generic "enable" one, where we can
enable the cpython freethreading builds. PyPy builds also have to be
explicitly enabled now, since they are going to be turned off by default
in cibw v3, we have historically supported them, and because a warning
is now issued if they are not explicitly listed in the "enable" option.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>