Check incoming values in yajl2_c backend
When events are sent to the yajl2_c backend, we assumed the value sent
by the user was a tuple of the correct length. This is usually correct,
but it's wrong to rely on it blindly.

This commit adds the necessary checks on input values sent to the send()
method of the *_basecoro generators of the yajl2_c backend. The
python-based backends already perform these checks by using the target
list assignment syntax (e.g., `a, b = c`) to receive any values sent to
them.

The checks in the yajl2_c backend are performed using a new ijson_unpack
utility function, which mimics the behaviour of target list assignments,
performing the same cardinality checks, and raising the same exceptions
when things go wrong. Note that the Python C API doesn't offer an
equivalent utility function, and the closest alternative (ensuring a
PyTuple, then invoking PyArg_UnpackTuple) doesn't yield the same error
messages and error types (TypeError v/s ValueError).

With this behaviour in place, we can now reliably test across all
backends that incorrect inputs raise certain exceptions.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>